requirements.txt
assemblyai==0.28.1
chromadb==0.5.0
Flask==3.0.3
Flask_Cors==4.0.1
Flask_SocketIO==5.3.6
langchain==0.2.4
langchain_community==0.2.4
langchain_core==0.2.10
langchain_openai==0.1.9
networkx==3.1
openai==1.34.0
opencv_python==4.7.0.68
openpyxl==3.1.5
pandas==2.0.3
psutil==6.0.0
psycopg2-binary==2.9.9
pydub==0.25.1
pypdf==4.2.0
pytesseract==0.3.10
python-dotenv==1.0.1
python-docx==1.1.2
unstructured==0.11.8
wcwidth==0.2.13
watchdog==4.0.1
wheel==0.43.0
yt_dlp==2024.7.1

